Update general.c
may cause pointer-to-int warning (different sizes) on i386 architecture

+++ src/rules/general.c
@@ -33,7 +33,7 @@ int general_seccomp_rules(struct config *_config) {
         return LOAD_SECCOMP_FAILED;
     }
     // add extra rule for execve
-    if (seccomp_rule_add(ctx, SCMP_ACT_KILL, SCMP_SYS(execve), 1, SCMP_A0(SCMP_CMP_NE, (scmp_datum_t)(_config->exe_path))) != 0) {
+    if (seccomp_rule_add(ctx, SCMP_ACT_KILL, SCMP_SYS(execve), 1, SCMP_A0(SCMP_CMP_NE, (scmp_datum_t)(size_t)(_config->exe_path))) != 0) {
         return LOAD_SECCOMP_FAILED;
     }
     // do not allow "w" and "rw" using open
